requirements.txt
aiohttp
PyYAML
requests
beautifulsoup4
